openbenchvue/requirements.txt
# Core instrument communication
pyvisa>=1.13.0
pyvisa-py>=0.7.0

# GUI framework
PyQt5>=5.15.0
PyQt5-sip>=12.11.0

# Data processing and analysis
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Configuration and serialization
pyyaml>=6.0

# Remote web interface
flask>=2.3.0
flask-cors>=4.0.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Logging and monitoring
colorlog>=6.7.0

# Optional: Excel export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Optional: Enhanced performance
numba>=0.57.0

# Development tools (optional)
pytest>=7.3.0
pytest-qt>=4.2.0
black>=23.3.0
pylint>=2.17.0
mypy>=1.3.0
